backend/requirements.txt
# FastAPI & Server
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Binance API
python-binance>=1.0.19

# Database
sqlalchemy>=2.0.25
aiomysql>=0.2.0  # MySQL support
asyncpg>=0.29.0  # PostgreSQL support (for Supabase)
aiosqlite>=0.19.0  # SQLite support (for data_collector)
alembic>=1.13.0

# AI/ML
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
torch>=2.2.0
faiss-cpu>=1.7.4

# Vector DB
google-generativeai>=0.5.0
transformers>=4.37.0

# WebSocket for real-time data
websockets>=12.0

# HTTP client
httpx>=0.27.0
aiohttp>=3.9.0

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
